tools/python/requirements.txt
numpy
pandas
scipy.io